Adding files to start chrome extension
Put all these files in a folder, that folder will be the one up upload to chrome://extensions. Once thats started I included instructions in the python folder to start the server. Once the server is started, and the extension is installed, once you change tabs or open new windows, they will print to the terminal.